+++ src/app/[variants]/(main)/_layout/Desktop/index.tsx
@@ -9,12 +9,12 @@ import { Flexbox } from 'react-layout-kit';
 
 import { isDesktop } from '@/const/version';
 import { BANNER_HEIGHT } from '@/features/AlertBanner/CloudBanner';
+import TitleBar, { TITLE_BAR_HEIGHT } from '@/features/ElectronTitlebar';
 import HotkeyHelperPanel from '@/features/HotkeyHelperPanel';
 import { usePlatform } from '@/hooks/usePlatform';
 import { featureFlagsSelectors, useServerConfigStore } from '@/store/serverConfig';
 import { HotkeyScopeEnum } from '@/types/hotkey';
 
-import TitleBar, { TITLE_BAR_HEIGHT } from './ElectronTitlebar';
 import RegisterHotkeys from './RegisterHotkeys';
 import SideBar from './SideBar';
